+++ src/components/index.ts
@@ -1,4 +1,5 @@
 // BEGIN_OF_AL_CLI_COMPONENTS
 export {default as ClientCounter} from './sample-name/ClientCounter';
 export {default as ClientPrefix} from './sample-name/ClientPrefix';
+export {default as ClientLaunchFirstAppBar} from './sample-name/ClientLaunchFirstAppBar';
 // END_OF_AL_CLI_COMPONENTS