+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.0] 2019-06-16
+- Update dependencies to
+    - @material-ui/core @ ^4.0.2
+    - create-react-app @ 3.0.1
+- Remove dependencies
+    - material-ui-pickers @ ^2.2.4
+
 ## [0.70.40] 2019-04-23
 - Integrate react-openfin-scripts@^0.40.20 to simplify the scaffolding process